feat : Added community review writing functionality and UI to the Review Detail page.
