+++ front/public/image.zip
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6076520882b5ca2646b4f32a63382a3d402cec8e1b86e66e3d02cf12453d198a
+size 120046884
