+++ front/tailwind.config.js
@@ -98,6 +98,17 @@ module.exports = {
         },
       },
     },
+
+    screens: {
+      "3sm": "320px",
+      "2sm": "450px",
+      sm: "625px",
+      md: "768px",
+      lg: "1024px",
+      xl: "1280px",
+      "2xl": "1536px",
+      "3xl": "1921px",
+    },
   },
 
   corePlugins: {
